requirements.txt
streamlit 
pandas 
python-dotenv 
schedule
sendgrid
langchain-groq
langchain_community
astrapy
streamlit-calendar
streamlit-option-menu
streamlit-lottie